# backend/cmd/api/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Backend API built with Golang using Clean Code architecture. \\nGithub: [https://github.com/shion0625/backend].",
    "title": "E-commerce Application Backend API",
    "contact": {
      "name": "For API Support",
      "email": "nikhilnarayanan623@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "basePath": "/api",
  "paths": {
    "/account": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for user to get all user details",
        "tags": ["User Profile"],
        "summary": "Get User Profile (User)",
        "operationId": "GetProfile",
        "responses": {
          "200": {
            "description": "Successfully retrieved user details"
          },
          "500": {
            "description": "Failed to retrieve user details",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for user to edit user details",
        "tags": ["User Profile"],
        "summary": "Edit profile (User)",
        "operationId": "UpdateProfile",
        "parameters": [
          {
            "description": "User details input",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.EditUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully profile updated",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "400": {
            "description": "Invalid inputs",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to update profile",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/account/address": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for user to get all user addresses",
        "tags": ["User Profile"],
        "summary": "Get all addresses (User)",
        "operationId": "GetAllAddresses",
        "responses": {
          "200": {
            "description": "successfully retrieved all user addresses",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "failed to show user addresses",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for user to update user address",
        "tags": ["User Profile"],
        "summary": "Update address (User)",
        "operationId": "UpdateAddress",
        "parameters": [
          {
            "description": "Address input",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.EditAddress"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfully addresses updated",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "400": {
            "description": "can't update the address",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for user to add a new address",
        "tags": ["User Profile"],
        "summary": "Add a new address (User)",
        "operationId": "SaveAddress",
        "parameters": [
          {
            "description": "Address input",
            "name": "inputs",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully address added",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "400": {
            "description": "invalid input",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to save address",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
...
    "/account/wishlist": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": ["User Profile"],
        "summary": "Get whish list product items (User)",
        "operationId": "GetWishList",
        "responses": {
          "200": {
            "description": "Successfully retrieved all product items in th wish list"
          },
          "500": {
            "description": "Failed to retrieve product items from the wish list"
          }
        }
      }
    },
    "/account/wishlist/{product_item_id}": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": ["User Profile"],
        "summary": "Add to whish list (User)",
        "operationId": "SaveToWishList",
        "parameters": [
          {
            "type": "integer",
            "description": "Product Item ID",
            "name": "product_item_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully product items added to whish list",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "400": {
            "description": "invalid input",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "409": {
            "description": "Product item already exist on wish list",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to add product item to wishlist",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": ["User Profile"],
        "summary": "Remove from whish list (User)",
        "operationId": "RemoveFromWishList",
        "parameters": [
          {
            "type": "integer",
            "description": "Product Item ID",
            "name": "product_item_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successfully removed product item from wishlist",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "400": {
            "description": "invalid input",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
...
    "/admin/stocks": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for admin to get all stocks",
        "tags": ["Admin Stock"],
        "summary": "Get all stocks (Admin)",
        "operationId": "GetAllStocks",
        "parameters": [
          {
            "type": "integer",
            "description": "Page Number",
            "name": "page_number",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "Count",
            "name": "count",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found all stocks",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "204": {
            "description": "No stocks found",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to Get all stocks",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      },
      "patch": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "API for admin to update stock details",
        "tags": ["Admin Stock"],
        "summary": "Update stocks (Admin)",
        "operationId": "UpdateStock",
        "parameters": [
          {
            "description": "Update stock details",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.UpdateStock"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated sock",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "400": {
            "description": "Failed to bind input",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          },
          "500": {
            "description": "Failed to update stock",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "request.Address": {
      "type": "object",
      "required": ["house", "land_mark", "name", "phone_number", "pincode"],
      "properties": {
        "area": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "house": {
          "type": "string"
        },
        "is_default": {
          "type": "boolean"
        },
        "land_mark": {
          "type": "string"
        },
        "name": {
          "type": "string",
          "maxLength": 50,
          "minLength": 2
        },
        "phone_number": {
          "type": "string",
          "maxLength": 10,
          "minLength": 10
        },
        "pincode": {
          "type": "integer"
        }
      }
    },
...
    "request.EditAddress": {
      "type": "object",
      "required": [
        "address_id",
        "house",
        "land_mark",
        "name",
        "phone_number",
        "pincode"
      ],
      "properties": {
        "address_id": {
          "type": "integer"
        },
        "area": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "house": {
          "type": "string"
        },
        "is_default": {
          "type": "boolean"
        },
        "land_mark": {
          "type": "string"
        },
        "name": {
          "type": "string",
          "maxLength": 50,
          "minLength": 2
        },
        "phone_number": {
          "type": "string",
          "maxLength": 10,
          "minLength": 10
        },
        "pincode": {
          "type": "integer"
        }
      }
    },
...
    "request.EditUser": {
      "type": "object",
      "required": [
        "age",
        "email",
        "first_name",
        "last_name",
        "phone",
        "user_name"
      ],
      "properties": {
        "age": {
          "type": "integer"
        },
        "confirm_password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "first_name": {
          "type": "string",
          "maxLength": 50,
          "minLength": 2
        },
        "last_name": {
          "type": "string",
          "maxLength": 50,
          "minLength": 1
        },
        "password": {
          "type": "string"
        },
        "phone": {
          "type": "string",
          "maxLength": 10,
          "minLength": 10
        },
        "user_name": {
          "type": "string",
          "maxLength": 15,
          "minLength": 3
        }
      }
    },
...
    "request.UpdateStock": {
      "type": "object",
      "properties": {
        "qty_to_add": {
          "type": "integer"
        },
        "sku": {
          "type": "string"
        }
      }
    },
...
    "response.Response": {
      "type": "object",
      "properties": {
        "data": {},
        "error": {},
        "message": {
          "type": "string"
        },
        "success": {
          "type": "boolean"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Add prefix of Bearer before  token Ex: \"Bearer token\"",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}